pos_disable_payment: [IMP] add restrictions on pos config
Before this commit, it was not possible to limit rights on a specific pos.
However, limiting actions to a special pos can help to establish a more
fine grained control. Example, when configuring an order-only station.

A sepcial (hard) requirement is, if attributes of the pos are rather connected
with the physical point (tablet) rather than the person using it.

One example is offline sequence management, where for example invoicing is only
permitted on audited/authorized devices.